addTransactionValid

+++ src/main/java/com/webmusic/controller/SingerController.java
@@ -8,6 +8,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
 
@@ -32,19 +33,22 @@ public class SingerController {
     }
 
     @PostMapping("/add")
-    public ResponseEntity<Singer> addSinger(@RequestBody Singer singer) {
+    public ResponseEntity<Singer> addSinger(@Valid @RequestBody Singer singer) {
         Singer addSinger = singerService.save(singer);
         return new ResponseEntity<>(addSinger, CREATED);
     }
 
     @PutMapping("/edit")
-    public ResponseEntity<Singer> editSinger(@RequestBody Singer singer) {
+    public ResponseEntity<Singer> editSinger(@Valid @RequestBody Singer singer) {
         Singer editSinger = singerService.save(singer);
         return new ResponseEntity<>(editSinger, OK);
     }
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> deleteSinger(@PathVariable Long id) {
+        if (id == null){
+            return new ResponseEntity<>(NOT_FOUND);
+        }
         singerService.delete(id);
         return new ResponseEntity<>(OK);
     }
